requirements.txt
flask
transformers
torch
pillow
